using single precision floats on functions.c

+++ mhps/lib/functions.c
@@ -8,37 +8,37 @@ float _conv(int N, float *a, float *time1, float *b1, float *b2,
   float norm1, norm2, normf, t, dt1, min, max;
 
   for (ci=0; ci<d; ci++){
-    b1[ci] = 0.0;
-    b2[ci] = 0.0;
+    b1[ci] = 0.0f;
+    b2[ci] = 0.0f;
   }
 
-  norm1 = 0.0;
-  norm2 = 0.0;
+  norm1 = 0.0f;
+  norm2 = 0.0f;
 
-  min = -6.0*c1;
-  max = 6.0*c1;
+  min = -6.0f*c1;
+  max = 6.0f*c1;
   
-  if (max < 6.0*dtmin){
-    min = -6.0*dtmin;
-    max = 6.0*dtmin;
+  if (max < 6.0f*dtmin){
+    min = -6.0f*dtmin;
+    max = 6.0f*dtmin;
   }
-  dt1 = (max-min)/1000.0;
+  dt1 = (max-min)/1000.0f;
 
-  norm1 = 0.0;
-  norm2 = 0.0;
+  norm1 = 0.0f;
+  norm2 = 0.0f;
   for (t=min; t<=max; t=t+dt1){
-    norm1 += pow(E, -t*t/(2*c1*c1));
-    norm2 += pow(E, -t*t/(2*c2*c2));
+    norm1 += powf(E, -t*t/(2*c1*c1));
+    norm2 += powf(E, -t*t/(2*c2*c2));
   }
   norm1 = norm1*dt1;
   norm2 = norm2*dt1;
 
-  normf = 0.0;
+  normf = 0.0f;
   for (t=min; t<=max; t=t+dt1){
-    normf += pow(
-        pow(E, -t*t/(2*c1*c1))/norm1 
-        - pow(E, -t*t/(2*c2*c2))/norm2, 
-        2.0);
+    normf += powf(
+        powf(E, -t*t/(2*c1*c1))/norm1 
+        - powf(E, -t*t/(2*c2*c2))/norm2, 
+        2.0f);
   }
   normf = normf*dt1;
 
@@ -46,10 +46,10 @@ float _conv(int N, float *a, float *time1, float *b1, float *b2,
     k = 0;
     for (cj=0; cj<d; cj++){
       dt1 = time1[cj] - time1[ci];
-      if (dt1>min && dt1<max && mask[cj]>0.0){
-          b1[ci] += a[cj]*pow(E, -dt1*dt1/(2*c1*c1))/norm1;
-          b2[ci] += a[cj]*pow(E, -dt1*dt1/(2*c2*c2))/norm2;
-          if (dt1>=min/3.0*2.0 && dt1<=max/3.0*2.0){
+      if (dt1>min && dt1<max && mask[cj]>0.0f){
+          b1[ci] += a[cj]*powf(E, -dt1*dt1/(2*c1*c1))/norm1;
+          b2[ci] += a[cj]*powf(E, -dt1*dt1/(2*c2*c2))/norm2;
+          if (dt1>=min/3.0f*2.0f && dt1<=max/3.0f*2.0f){
               k++;
           }
       }
@@ -57,8 +57,8 @@ float _conv(int N, float *a, float *time1, float *b1, float *b2,
     b1[ci] *= mask[ci];
     b2[ci] *= mask[ci];
     if (k<=3){
-      b1[ci] = -1.0;
-      b2[ci] = -1.0;
+      b1[ci] = -1.0f;
+      b2[ci] = -1.0f;
     }
   }
   return (normf);
@@ -72,25 +72,25 @@ void _statistics(int N, float *flux, float* mag, float *magerr, float* time, flo
   float kr, sigma1, sigma2, norm, PN, Ik, fluxerr, Ik2;
   int imax, k;
 
-  kr = 1.0/t1;
-  sigma1 = 0.225/kr/sqrt(1.0+epsilon);
-  sigma2 = (1.0+epsilon)*sigma1;
+  kr = 1.0f/t1;
+  sigma1 = 0.225f/kr/sqrtf(1.0f+epsilon);
+  sigma2 = (1.0f+epsilon)*sigma1;
 
   imax = N;
   norm = _conv(N, mask, time, maskG1, maskG2, mask, sigma1, sigma2, imax, dt);
   norm = _conv(N, flux, time, rateG1, rateG2, mask, sigma1, sigma2, imax, dt);
 
   k = 0;
-  Ik2 = 0.0;
+  Ik2 = 0.0f;
   *nonzero = 0;
-  PN = 0.0;
+  PN = 0.0f;
 
   while (k<imax){
-      if (maskG1[k]>0.0){
+      if (maskG1[k] > 0.0f){
           Ik = rateG1[k]/maskG1[k]-rateG2[k]/maskG2[k];
           Ik2 += Ik*Ik;
           (*nonzero)++;
-          fluxerr = log(10.0)*pow(10.0,-0.4*(mag[k]-mag0))*0.4*magerr[k];
+          fluxerr = logf(10.0f)*powf(10.0f, -0.4f*(mag[k]-mag0))*0.4f*magerr[k];
           PN += fluxerr;
           }
       k++;
@@ -103,19 +103,19 @@ void _statistics(int N, float *flux, float* mag, float *magerr, float* time, flo
   Ik2 -= PN;
   *Ik2_low_freq = Ik2;
 
-  kr = 1.0/t2;
-  sigma1 = 0.225/kr/sqrt(1.0+epsilon);
-  sigma2 = (1.0+epsilon)*sigma1;
+  kr = 1.0f/t2;
+  sigma1 = 0.225f/kr/sqrtf(1.0f+epsilon);
+  sigma2 = (1.0f+epsilon)*sigma1;
 
 
   norm = _conv(N, flux, time, rateG1, rateG2, mask, sigma1, sigma2, imax, dt);
   norm = _conv(N, mask, time, maskG1, maskG2, mask, sigma1, sigma2, imax, dt);
 
   k = 0;
-  Ik2 = 0.0;
+  Ik2 = 0.0f;
   *nonzero = 0;
-  while (k<imax){
-      if (maskG1[k]>0.0){
+  while (k < imax){
+      if (maskG1[k] > 0.0f){
           Ik = rateG1[k]/maskG1[k]-rateG2[k]/maskG2[k];
           Ik2 += Ik*Ik;
           (*nonzero)++;
@@ -143,8 +143,8 @@ void _flux_statistics(
   float kr, sigma1, sigma2, norm, PN, Ik, Ik2;
   int imax, k;
 
-  kr = 1.0/t1;
-  sigma1 = 0.225/kr/sqrt(1.0+epsilon);
+  kr = 1.0f/t1;
+  sigma1 = 0.225f/kr/sqrtf(1.0f+epsilon);
   sigma2 = (1.0+epsilon)*sigma1;
 
   imax = N;
@@ -152,12 +152,12 @@ void _flux_statistics(
   norm = _conv(N, flux, time, rateG1, rateG2, mask, sigma1, sigma2, imax, dt);
 
   k = 0;
-  Ik2 = 0.0;
+  Ik2 = 0.0f;
   *nonzero = 0;
-  PN = 0.0;
+  PN = 0.0f;
 
   while (k < imax){
-      if (maskG1[k] > 0.0){
+      if (maskG1[k] > 0.0f){
           Ik = rateG1[k]/maskG1[k]-rateG2[k]/maskG2[k];
           Ik2 += Ik*Ik;
           (*nonzero)++;
@@ -173,19 +173,19 @@ void _flux_statistics(
   Ik2 -= PN;
   *Ik2_low_freq = Ik2;
 
-  kr = 1.0/t2;
-  sigma1 = 0.225/kr/sqrt(1.0+epsilon);
-  sigma2 = (1.0+epsilon)*sigma1;
+  kr = 1.0f/t2;
+  sigma1 = 0.225f/kr/sqrtf(1.0f+epsilon);
+  sigma2 = (1.0f+epsilon)*sigma1;
 
 
   norm = _conv(N, flux, time, rateG1, rateG2, mask, sigma1, sigma2, imax, dt);
   norm = _conv(N, mask, time, maskG1, maskG2, mask, sigma1, sigma2, imax, dt);
 
   k = 0;
-  Ik2 = 0.0;
+  Ik2 = 0.0f;
   *nonzero = 0;
-  while (k<imax){
-      if (maskG1[k]>0.0){
+  while (k < imax){
+      if (maskG1[k] > 0.0f){
           Ik = rateG1[k]/maskG1[k]-rateG2[k]/maskG2[k];
           Ik2 += Ik*Ik;
           (*nonzero)++;
@@ -211,15 +211,15 @@ long _sigma_clip(float *flux, float *flag, float mean, float var, float *new_mea
   float new_mean_value, new_var_value, new_new_mean_value, new_sigma;
 
   new_num = 0;
-  flux2 = new_mean_value = 0.0;
-  sigma = sqrt(var);
+  flux2 = new_mean_value = 0.0f;
+  sigma = sqrtf(var);
 
   for(i=0;i<num;i++){
-    if(flux[i]>(mean+3.0*sigma) ||flux[i]<(mean-3.0*sigma))
-      flag[i]=10.0;
+    if(flux[i]>(mean+3.0f*sigma) ||flux[i]<(mean-3.0f*sigma))
+      flag[i]=10.0f;
   }
   for(i=0;i<num;i++){
-    if (flag[i]<1.0){
+    if (flag[i]<1.0f){
       new_mean_value+=flux[i];
       flux2+=flux[i]*flux[i];
       new_num++;
@@ -231,18 +231,18 @@ long _sigma_clip(float *flux, float *flag, float mean, float var, float *new_mea
     new_var_value=flux2-new_mean_value*new_mean_value;
   }
 
-  if((new_var_value-var)*(new_var_value-var)>(new_var_value/10.0)*(new_var_value/10.0)){
-      new_sigma=sqrt(new_var_value);
+  if((new_var_value-var)*(new_var_value-var)>(new_var_value/10.0f)*(new_var_value/10.0f)){
+      new_sigma=sqrtf(new_var_value);
       new_new_num=0;
-      flux2=0.0;
-      new_new_mean_value=0.0;
+      flux2=0.0f;
+      new_new_mean_value=0.0f;
 
-      for(i=0;i<num;i++){
-	if(flux[i]>(new_mean_value+3.0*new_sigma) ||flux[i]<(new_mean_value-3.0*new_sigma))
-	  flag[i]=10.0;
+      for(i=0; i<num; i++){
+	    if(flux[i]>(new_mean_value+3.0f*new_sigma) ||flux[i]<(new_mean_value-3.0f*new_sigma))
+	    flag[i]=10.0f;
      }
       for(i=0;i<num;i++){
-	if (flag[i]<1.0){
+	if (flag[i]<1.0f){
 	  new_new_mean_value+=flux[i];
 	  flux2+=flux[i]*flux[i];
 	  new_new_num++;
@@ -250,13 +250,13 @@ long _sigma_clip(float *flux, float *flag, float mean, float var, float *new_mea
       }
       new_num=new_new_num;
       if (new_num>0){
-	new_mean_value=new_new_mean_value/(float)new_num;
-	flux2/=(float)new_num;
-	new_var_value=flux2-new_mean_value*new_mean_value;
+	new_mean_value = new_new_mean_value/(float)new_num;
+	flux2 /= (float)new_num;
+	new_var_value = flux2-new_mean_value*new_mean_value;
       }
     }
-    *new_mean=new_mean_value;
-    *new_var=new_var_value;
+    *new_mean = new_mean_value;
+    *new_var = new_var_value;
 
 
   return(new_num);
